TeaMachine is 25 days old today, time to release!
TeaMachine is a unique direct threaded virtual machine with .7 nanosecond cycle time. Not a typo. TeaMachine's superpower is a powerful key value store (Shardmap) built-in as a set of micro ops on an infinite set of database registers, each of which is a scalable key value store. Very early days for TeaMachine, it mostly consists of missing features, truth be told. But the big ones are already there, so jump on board and enjoy what is shaping up to be a great ride.