Call tpcb_main from teamachine with command line strings
This is a proof of concept that teamachine can call main-like extern
"C" functions with signature int (fn)(int argc, char *argv[]) without
too much fuss or muss. The tpcb_main function was changed to return
instead of call exit(), which is probably better practice anyway. Two
fake strings were put on the stack to make create the fiction the call
is coming from a C main, which fiction seems rather well preserved.

This is really just a stunt, but it gives you and idea what teamachine
can do already. The native call interface is very powerful and also
efficient. This is just one example of what you can do with it.